msm8998-common: vendor: data: Reconfigure MSM8998 RMNET
The configuration provided by kholk causes page allocation failures
due to the enablement of data aggregation of upload packets which
our platform does not support. Restore the stock msm8998 upload packets
configuration.

Furthermore, some of the newly added configurations are outdated compared
to our kernel's RMNET actual capabilities. Update them to fully make use of
the new RMNET drivers by basing the configuration on msmnile.

TCP MTU probing should be returned to what OnePlus' recommends as
it is not a requirement for supporting the new RMNET.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>